rename ShipParameters widget folder
