+++ lib/presentation/strength/widgets/ship_parameters/confirmation_dialog.dart
